@@ -29,7 +29,6 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.13.10",
-    "@syncfusion/ej2-layouts": "*",
     "@syncfusion/ej2-react-calendars": "*",
     "@syncfusion/ej2-react-charts": "*",
     "@syncfusion/ej2-react-dropdowns": "*",
@@ -38,8 +37,8 @@
     "@syncfusion/ej2-react-progressbar": "*",
     "css-loader": "^6.7.1",
     "file-loader": "^6.2.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^19.2.3",
+    "react-dom": "^19.2.3",
     "style-loader": "^3.3.1",
     "url-loader": "^4.1.1"
   }
