@@ -35,7 +35,7 @@
     "@syncfusion/ej2-react-calendars": "*",
     "@syncfusion/ej2-react-charts": "*",
     "@syncfusion/ej2-react-popups": "*",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react": "^19.2.3",
+    "react-dom": "^19.2.3"
   }
 }
