@@ -33,7 +33,7 @@
     "@babel/runtime": "^7.13.10",
     "@syncfusion/ej2-react-calendars": "*",
     "file-loader": "^6.2.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react": "^19.2.3",
+    "react-dom": "^19.2.3"
   }
 }
